requirements.txt
Django==1.10.5
django-dbbackup==3.1.3
django-environ==0.4.1
gunicorn==19.6.0
ipython==5.1.0
mysqlclient==1.3.9
psycopg2==2.6.2
requests==2.12.4
PyJWT==1.4.2
